+++ as-pect.config.js
@@ -18,7 +18,8 @@ module.exports = {
     /** To enable wat file output, use the following flag. The filename is ignored, but required by the compiler. */
     // "--textFile": ["output.wat"],
     /** To select an appropriate runtime, use the --runtime compiler flag. */
-    "--runtime": ["stub"] // Acceptable values are: full, half, stub (arena), and none
+    "--runtime": ["stub"], // Acceptable values are: full, half, stub (arena), and none
+    "--baseDir": process.cwd()
   },
   /**
    * A set of regexp that will disclude source files from testing.
