+++ assembly/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../node_modules/assemblyscript/std/assembly.json",
+  "include": [
+    "**/*.ts"
+  ]
+}
