+++ assembly/types.ts
@@ -1,0 +1,20 @@
+
+
+export const DEFAULT_SCRATCH_BUFFER_SIZE: usize = 1 << 16;
+
+
+export const enum DataTypeIndex {
+    DATA_TYPE_ORIGINATOR_ACCOUNT_ID = 1,
+    DATA_TYPE_CURRENT_ACCOUNT_ID = 2,
+    DATA_TYPE_STORAGE = 3,
+    DATA_TYPE_INPUT = 4,
+    DATA_TYPE_RESULT = 5,
+    DATA_TYPE_STORAGE_ITER = 6,   
+}
+
+export const DATA_TYPE_ORIGINATOR_ACCOUNT_ID = DataTypeIndex.DATA_TYPE_ORIGINATOR_ACCOUNT_ID;
+export const DATA_TYPE_CURRENT_ACCOUNT_ID = DataTypeIndex.DATA_TYPE_CURRENT_ACCOUNT_ID;
+export const DATA_TYPE_STORAGE = DataTypeIndex.DATA_TYPE_STORAGE;
+export const DATA_TYPE_INPUT = DataTypeIndex.DATA_TYPE_INPUT;
+export const DATA_TYPE_RESULT = DataTypeIndex.DATA_TYPE_RESULT;
+export const DATA_TYPE_STORAGE_ITER = DataTypeIndex.DATA_TYPE_STORAGE_ITER;
