+++ package.json
@@ -2,16 +2,17 @@
   "name": "near-runtime-ts",
   "version": "0.0.2",
   "description": "Near smart contract runtime library for typescript",
-  "main": "index.js",
+  "types": "assembly/index.ts",
+  "ascMain": "assembly/index.ts",
   "scripts": {
     "doc": "rm -rf apidoc; mkdir apidoc; node_modules/.bin/typedoc near.ts --theme markdown --ignoreCompilerErrors --excludePrivate --excludeProtected --excludeExternals --out apidoc/",
     "test": "asp --types; asp"
   },
   "author": "",
   "license": "MIT",
-  "dependencies": {},
   "devDependencies": {
-    "as-pect": "github:jtenner/as-pect",
+    "@as-pect/cli": "^2.3.1",
+    "assemblyscript": "github:AssemblyScript/assemblyscript",
     "bignum": "MaxGraey/bignum.wasm",
     "typedoc": "^0.14.2",
     "typedoc-plugin-markdown": "^1.2.1"
